+++ template/tasks.py
@@ -7,6 +7,7 @@ import os
 import shutil
 import urllib.parse
 import githubkit
+import requests
 from azure.devops.connection import Connection
 from azure.devops.exceptions import AzureDevOpsServiceError
 from msrest.authentication import BasicAuthentication
@@ -67,7 +68,7 @@ def repo_create_azdo(c, answers_json):
     }
     print("[cyan]Creating repo in Azure DevOps...[/cyan]")
     git_client.create_repository(repo_data, project=answers["azdo_project"])
-    print("[bold green]*** 'repo-create-github' task end ***[/bold green]")
+    print("[bold green]*** 'repo-create-azdo' task end ***[/bold green]")
 
 @task
 def repo_settings_github(c, answers_json):
@@ -197,6 +198,45 @@ def initialize_and_commit(c, answers_json):
     print("[cyan]Pushing to remote...[/cyan]")
     c.run(f"git push -u origin --all")
     print("[bold green]*** 'initialize_and_commit' task end ***[/bold green]")
+
+@task
+def pipelines_create_azdo(c, answers_json):
+    """Register pipeliens for an Azure DevOps repo"""
+    # Note that pipeline creation with the python module straight-up doesn't
+    # work, so gotta do this the old-fashioned way.
+    # See https://github.com/microsoft/azure-devops-python-api/issues/432
+    print("[bold green]*** 'pipelines_create_azdo' task start ***[/bold green]")
+    answers = json.loads(answers_json)
+    with open("token.json") as token_file:
+        token = json.loads(token_file.read())["token"]
+
+    print("[cyan]Authenticating to Azure DevOps...[/cyan]")
+    credentials = BasicAuthentication('', token)
+    connection = Connection(base_url=f"https://dev.azure.com/{answers['azdo_org']}", creds=credentials)
+    git_client = connection.clients_v5_1.get_git_client()
+    repo_return = git_client.get_repositories(project=answers["azdo_project"])
+    matching_repo_object = [x for x in repo_return if x.name == "testrepo1"][0]
+    repo_id = matching_repo_object.id
+
+    release_pipeline_data = {
+        "configuration": {
+            "type": "yaml",
+            "path": ".azurepipelines/azure-pipelines-release.yml",
+            "repository": {
+                "id": repo_id,
+                "type": "azureReposGit"
+            }
+        },
+        "name": answers["repo_name"],
+        "folder": "\\"
+    }
+    print("[cyan]Creating release pipeline in Azure DevOps...[/cyan]")
+    headers = {
+        "Content-Type": "application/json",
+        "Accept": "application/json"
+    }
+    requests.post(f"https://dev.azure.com/{answers['azdo_org']}/{answers['azdo_project']}/_apis/pipelines?api-version=7.1-preview.1", data=json.dumps(release_pipeline_data), auth=('', token), headers=headers)
+    print("[bold green]*** 'pipelines_create_azdo' task end ***[/bold green]")
 
 @task
 def delete_files(c):