+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## Added
+- Add Azure DevOps template release pipeline
+- Add Azure DevOps pipeline creation
 
 ## [0.1.14] - 2024-04-02
 ## Fixed